Refactor all theme file classes to inherit from ThemeFile
Which implement all comment methods used by Analyzer

Fixes #339
