+++ lib/theme_check.rb
@@ -4,6 +4,7 @@ require "liquid"
 require_relative "theme_check/version"
 require_relative "theme_check/bug"
 require_relative "theme_check/exceptions"
+require_relative "theme_check/theme_file"
 require_relative "theme_check/analyzer"
 require_relative "theme_check/check"
 require_relative "theme_check/checks_tracking"
