+++ lib/theme_check/asset_file.rb
@@ -1,27 +1,15 @@
 # frozen_string_literal: true
-require "pathname"
 require "zlib"
 
 module ThemeCheck
-  class AssetFile
+  class AssetFile < ThemeFile
     def initialize(relative_path, storage)
-      @relative_path = relative_path
-      @storage = storage
+      super
       @loaded = false
       @content = nil
     end
 
-    def path
-      @storage.path(@relative_path)
-    end
-
-    def relative_path
-      @relative_pathname ||= Pathname.new(@relative_path)
-    end
-
-    def content
-      @content ||= @storage.read(@relative_path)
-    end
+    alias_method :content, :source
 
     def gzipped_size
       @gzipped_size ||= Zlib.gzip(content).bytesize
