+++ lib/theme_check/json_file.rb
@@ -1,27 +1,13 @@
 # frozen_string_literal: true
 require "json"
-require "pathname"
 
 module ThemeCheck
-  class JsonFile
+  class JsonFile < ThemeFile
     def initialize(relative_path, storage)
-      @relative_path = relative_path
-      @storage = storage
+      super
       @loaded = false
       @content = nil
       @parser_error = nil
-    end
-
-    def path
-      @storage.path(@relative_path)
-    end
-
-    def relative_path
-      @relative_pathname ||= Pathname.new(@relative_path)
-    end
-
-    def source
-      @source ||= @storage.read(@relative_path)
     end
 
     def content
@@ -34,22 +20,9 @@ module ThemeCheck
       @parser_error
     end
 
-    def name
-      relative_path.sub_ext('').to_s
-    end
-
     def json?
       true
     end
-
-    def liquid?
-      false
-    end
-
-    def ==(other)
-      other.is_a?(JsonFile) && relative_path == other.relative_path
-    end
-    alias_method :eql?, :==
 
     private
