+++ lib/theme_check/template.rb
@@ -1,39 +1,13 @@
 # frozen_string_literal: true
-require "pathname"
 
 module ThemeCheck
-  class Template
-    def initialize(relative_path, storage)
-      @storage = storage
-      @relative_path = relative_path
-    end
-
-    def path
-      @storage.path(@relative_path)
-    end
-
-    def relative_path
-      @relative_pathname ||= Pathname.new(@relative_path)
-    end
-
-    def source
-      @source ||= @storage.read(@relative_path)
-    end
-
+  class Template < ThemeFile
     def write
       content = updated_content
       if source != content
         @storage.write(@relative_path, content)
         @source = content
       end
-    end
-
-    def name
-      relative_path.sub_ext('').to_s
-    end
-
-    def json?
-      false
     end
 
     def liquid?
@@ -87,11 +61,6 @@ module ThemeCheck
     def root
       parse.root
     end
-
-    def ==(other)
-      other.is_a?(Template) && relative_path == other.relative_path
-    end
-    alias_method :eql?, :==
 
     def self.parse(source)
       Tags.register_tags!
